Repair JalaliMoment

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -9,14 +10,14 @@ import { MessageService } from '../_services/message.service';
   styleUrls: ['./messages.component.css']
 })
 export class MessagesComponent implements OnInit {
-  messages: Message[] | null = [];
+  messages: Message[] = [];
   pagination: Pagination;
   container = 'Unread';
   pageNumber = 1;
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,8 +33,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages?.splice(this.messages.findIndex(m => m.id === id), 1);
+    // this.messageService.deleteMessage(id).subscribe(() => {
+    //   this.messages?.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('توجه !', 'آیا از حذف اطمینان داری ؟').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages?.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
   }
